Add Validation Attributes, simple Email Service

+++ BlazorServerWebUI/Models/OrderViewModel.cs
@@ -10,7 +10,7 @@
    public class OrderDetailViewModel
     {
         public int ProductId { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int Quantity { get; set; }
+        public double UnitPrice { get; set; }
     }
 }
